feat: go full waku
