+++ src/internal/vite-plugins.ts
@@ -98,36 +98,6 @@ export function langWatcher(config: Config.Config): PluginOption {
   }
 }
 
-export function rawMarkdown(config: Config.Config): PluginOption {
-  let viteConfig: ResolvedConfig
-
-  return {
-    name: 'vocs:raw-markdown',
-    enforce: 'pre',
-    configResolved(resolvedConfig) {
-      viteConfig = resolvedConfig
-    },
-    async generateBundle() {
-      const pagesDir = path.resolve(viteConfig.root, config.srcDir, config.pagesDir)
-      const pages = await Array.fromAsync(fs.glob(`${pagesDir}/**/*.{md,mdx}`))
-
-      for (const page of pages) {
-        const content = await fs.readFile(page, 'utf-8')
-        const relativePath = page.replace(pagesDir, '').replace(/\.mdx?$/, '.md')
-        const outputPath = relativePath.endsWith('/index.md')
-          ? relativePath.replace('/index.md', '.md')
-          : relativePath
-
-        this.emitFile({
-          type: 'asset',
-          fileName: outputPath.slice(1),
-          source: content,
-        })
-      }
-    },
-  }
-}
-
 export function llms(config: Config.Config): PluginOption {
   const { description, title } = config
   let viteConfig: ResolvedConfig
